platform: mark static nla_policy variables as const
These static variables really never be modified.
Mark them as const, which allows the linker to mark them as
read-only.

The problem is libnl3's API, which has these parameters
not as const. Add a workaround for that. Clearly libnl3 is
not gonna modify the policy, that the API was fixed too [1]

[1] https://github.com/thom311/libnl/commit/b4802a17a7655bfeee3e9c06e649d30b96dbad3b