core: don't close input fd in nm_utils_fd_get_contents()
The function should not close the input file descriptor; however
fdopen() associates the fd to the new stream so that when the stream
is closed, the fd is too. The result is a double close() and the
second call can in certain cases affect a wrong fd.

Use a duplicate fd for the stream.

Fixes: 1d9bdad1dfd33601288de5510ff466acbe2c4b08

https://bugzilla.redhat.com/show_bug.cgi?id=1451236
(cherry picked from commit 597072296a19be051ea70b017fd19505ebee6530)