build: disable lcov version check
Ideally we should switch to the standard macro AX_CODE_COVERAGE for
code coverage, but infortunately it's not supported on older distros
(e.g. on Travis).

Keep using the current macro, but remove the lcov version check as
it's not clear why it's needed in the first place (see [1]). The
latest version of lcov is now 1.13.

[1] http://git.savannah.gnu.org/gitweb/?p=autoconf-archive.git;a=commit;h=dedd290bdb647704d74c13992bd04275d9cb648e