device: show better logging message when setting MTU fails
Setting the MTU might fail when the underlying device's MTU
is not set.

Detect that case, and log a better warning message.

Unfortunately, it's tricky to detect whether this is a complete
failure, or whether we will later try again to change the MTU.
So, we log a failure, altough later we might fix it. It would
be better not to warn about non-errors.