device: don't configure automatic default-routes if an explicit gateway is set
Since commit 5c299454b49b165f645c25fd3e083c0bb747ad91 we can configure
multiple default-routes.

That is especially useful with IPv6 to configure multiple routers.
It will also be useful, once we allow configuring manual default-routes,
like regular static routes.

However the problem is, that the default-route for the manual gateway
and the gateway from DHCP both get the same metric. So it's undefined
which route is used. To avoid that problem, and to restore previous
behavior, don't accept any default-routes if a gateway is set.

Fixes: 5c299454b49b165f645c25fd3e083c0bb747ad91