settings: keep connection alive in delete_auth_cb()
Before commit 3ecb57fdc468 ("settings: get rid of callback arguments
for nm_settings_connection_delete()") audit logging was done in the
callback, which was run by nm_settings_connection_delete() with a
reference to the connection to keep it alive. Now we have to keep an
extra reference to ensure it doesn't go away.

Fixes: 3ecb57fdc46833f2377fa86e045cd0f8461eaab9