ifcfg-rh: sort bond options when writing a connection
Bond options are stored in a hash table and the order in which they
are returned by the API is not guaranteed. Sort them alphabetically so
that a connection will always be written in the same way, even if the
internal implementation of the hash table or the hashing function
changes, as it did in commit a6be2f4aa907 ("all: use nm_str_hash()
instead of g_str_hash()").

(cherry picked from commit 9a631a068e21fc193d81f1c421a0350c8aab52a6)