policy: avoid false positives for checking for autoactivation
We want to only check for autoconnect all, if something happend that
makes it possible that we can autoconnect now (while we couldn't
previously).

It's not a real problem to check more often then strictly necessary.
But add a check to rule out a few false-positives to avoid the
overhead of checking all devices for autoconnect.

(cherry picked from commit f0147a9c476e49aad373591f7ef1029b683cd560)