tui: fix creation of open Wi-Fi connections
Commit 6a4af482f02c ("nmtui: always create ethernet settings for VLAN
and wireless security for wifi.") changed nmtui to always add the
wireless security setting to the new connection, but without
initializing it. This leads to a crash that was fixed in 40fcf67a8415
("tui: fix crash creating Wi-Fi connection").

There is an additional bug: connections without authentication can't
be saved because the wireless security setting has uninitialized
fields.

To fix this, revert both patches (the first partially) because the
previous code did the right thing as it added the setting only when
needed.

Fixes: 6a4af482f02c4342c69cd38dd46c078aeaf60d0d

https://bugzilla.redhat.com/show_bug.cgi?id=1518167
(cherry picked from commit fead82f4192c24eb017eac405f98ccb02402453f)